requirements.txt
flask>=3.1.2
flask-sqlalchemy>=3.1.1
pillow>=12.1.0
psycopg2-binary>=2.9.11
reportlab>=4.4.9
sqlalchemy>=2.0.46
qrcode
werkzeug